requirements.txt
joblib==0.16.0
scikit-learn==0.23.1
pandas==1.0.5
streamlit==0.73.1
Pillow==8.0.1